pyproject.toml
[project]
name = "foodorder"
version = "0.1.0"
description = "Telegram bot that automates Instacart grocery ordering"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=22.0",
    "playwright>=1.40",
    "anthropic>=0.40.0",
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
foodorder = "src.main:main"
